chore: update gatsby to v3
