+++ website/content/blog/netlify-cms-2-0-launches-with-bitbucket-support-and-a-new-monorepo-architecture.md
@@ -7,7 +7,7 @@ description: >-
   Announcing the release of Netlify CMS v2.0, with new Bitbucket support and an
   improved project architecture designed to ease contribution and the extension
   of features.
-date: '2018-07-26'
+date: 2018-07-26T12:00:00.000Z
 ---
 Today we’re releasing Netlify CMS 2.0, which adds support for using Bitbucket as a backend.
