+++ website/package.json
@@ -14,30 +14,31 @@
   "homepage": "https://www.netlifycms.org/",
   "license": "MIT",
   "dependencies": {
+    "@emotion/cache": "^10.0.29",
     "@emotion/core": "^10.0.35",
     "@emotion/styled": "^10.0.27",
     "dayjs": "^1.8.23",
     "emotion-theming": "^10.0.27",
-    "gatsby": "2.32.8",
-    "gatsby-plugin-catch-links": "2.10.0",
+    "gatsby": "3.0.1",
+    "gatsby-plugin-catch-links": "3.0.0",
     "gatsby-plugin-emotion": "^4.2.0",
-    "gatsby-plugin-manifest": "2.12.1",
-    "gatsby-plugin-netlify-cms": "^4.2.0",
-    "gatsby-plugin-react-helmet": "3.10.0",
-    "gatsby-remark-autolink-headers": "2.11.0",
+    "gatsby-plugin-manifest": "3.0.0",
+    "gatsby-plugin-netlify-cms": "^5.0.0",
+    "gatsby-plugin-react-helmet": "4.0.0",
+    "gatsby-remark-autolink-headers": "3.0.0",
     "gatsby-remark-external-links": "^0.0.4",
-    "gatsby-remark-prismjs": "3.13.0",
-    "gatsby-source-filesystem": "2.11.1",
-    "gatsby-transformer-json": "2.11.0",
-    "gatsby-transformer-remark": "2.16.1",
-    "gatsby-transformer-yaml": "2.11.0",
+    "gatsby-remark-prismjs": "4.0.0",
+    "gatsby-source-filesystem": "3.0.0",
+    "gatsby-transformer-json": "3.0.0",
+    "gatsby-transformer-remark": "3.0.0",
+    "gatsby-transformer-yaml": "3.0.0",
     "js-yaml": "^4.0.0",
     "lodash": "^4.17.15",
     "moment": "^2.24.0",
     "netlify-cms-app": "^2.14.26",
     "prismjs": "^1.21.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
     "react-github-btn": "^1.1.1",
     "react-helmet": "^6.0.0",
     "react-markdown": "^5.0.0",
@@ -45,13 +46,10 @@
   },
   "devDependencies": {
     "babel-plugin-prismjs": "^2.0.1",
-    "babel-preset-gatsby": "^0.12.0",
+    "babel-preset-gatsby": "^1.0.0",
     "eslint": "^7.4.0",
     "eslint-plugin-import": "^2.20.1",
-    "markdownlint-cli": "^0.26.0"
-  },
-  "resolutions": {
-    "mini-css-extract-plugin": "0.11.2"
+    "markdownlint-cli": "^0.27.1"
   },
   "private": true
 }
