+++ website/src/cms/cms.js
@@ -4,9 +4,9 @@ import dayjs from 'dayjs';
 import Prism from 'prismjs';
 import { CacheProvider } from '@emotion/core';
 import createCache from '@emotion/cache';
-import { BlogPostTemplate } from '../templates/blog-post';
+import BlogPostTemplate from '../components/blog-post-template';
 import { LayoutTemplate as Layout } from '../components/layout';
-import { DocsTemplate } from '../templates/doc-page';
+import DocsTemplate from '../components/docs-template';
 import WidgetDoc from '../components/widget-doc';
 import WhatsNew from '../components/whats-new';
 import Notification from '../components/notification';
