+++ website/src/templates/blog-post.js
@@ -2,34 +2,10 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import { graphql } from 'gatsby';
 import { trimStart, trimEnd } from 'lodash';
-import { css } from '@emotion/core';
 
 import TwitterMeta from '../components/twitter-meta';
 import Layout from '../components/layout';
-import Container from '../components/container';
-import Markdown from '../components/markdown';
-import MetaInfo from '../components/meta-info';
-import Page from '../components/page';
-
-export function BlogPostTemplate({ title, author, date, body, html }) {
-  return (
-    <Container size="sm">
-      <Page as="article">
-        <h1
-          css={css`
-            margin-bottom: 0;
-          `}
-        >
-          {title}
-        </h1>
-        <MetaInfo>
-          by {author} on {date}
-        </MetaInfo>
-        <Markdown body={body} html={html} />
-      </Page>
-    </Container>
-  );
-}
+import BlogPostTemplate from '../components/blog-post-template';
 
 function BlogPost({ data }) {
   const { html, frontmatter } = data.markdownRemark;
